moved constants to the functions that use them
