+++ Widgets.go
@@ -106,6 +106,26 @@ func ProgressBar(fraction float32) {
 	ProgressBarV(fraction, Vec2{X: -1, Y: 0}, "")
 }
 
+const (
+	// ComboFlagNone default = 0
+	ComboFlagNone = 0
+	// ComboFlagPopupAlignLeft aligns the popup toward the left by default.
+	ComboFlagPopupAlignLeft = 1 << 0
+	// ComboFlagHeightSmall has max ~4 items visible.
+	// Tip: If you want your combo popup to be a specific size you can use SetNextWindowSizeConstraints() prior to calling BeginCombo().
+	ComboFlagHeightSmall = 1 << 1
+	// ComboFlagHeightRegular has max ~8 items visible (default).
+	ComboFlagHeightRegular = 1 << 2
+	// ComboFlagHeightLarge has max ~20 items visible.
+	ComboFlagHeightLarge = 1 << 3
+	// ComboFlagHeightLargest has as many fitting items as possible.
+	ComboFlagHeightLargest = 1 << 4
+	// ComboFlagNoArrowButton displays on the preview box without the square arrow button.
+	ComboFlagNoArrowButton = 1 << 5
+	// ComboFlagNoPreview displays only a square arrow button.
+	ComboFlagNoPreview = 1 << 6
+)
+
 // BeginComboV creates a combo box with complete control over the content to the user.
 // Call EndCombo() if this function returns true.
 // flags are the ComboFlags to apply.
@@ -240,6 +260,52 @@ func VSliderInt(label string, size Vec2, value *int32, min, max int32) bool {
 	return VSliderIntV(label, size, value, min, max, "%d")
 }
 
+const (
+	// InputTextFlagsNone sets everything default.
+	InputTextFlagsNone = 0
+	// InputTextFlagsCharsDecimal allows 0123456789.+-
+	InputTextFlagsCharsDecimal = 1 << 0
+	// InputTextFlagsCharsHexadecimal allow 0123456789ABCDEFabcdef
+	InputTextFlagsCharsHexadecimal = 1 << 1
+	// InputTextFlagsCharsUppercase turns a..z into A..Z.
+	InputTextFlagsCharsUppercase = 1 << 2
+	// InputTextFlagsCharsNoBlank filters out spaces, tabs.
+	InputTextFlagsCharsNoBlank = 1 << 3
+	// InputTextFlagsAutoSelectAll selects entire text when first taking mouse focus.
+	InputTextFlagsAutoSelectAll = 1 << 4
+	// InputTextFlagsEnterReturnsTrue returns 'true' when Enter is pressed (as opposed to when the value was modified).
+	InputTextFlagsEnterReturnsTrue = 1 << 5
+	// InputTextFlagsCallbackCompletion for callback on pressing TAB (for completion handling).
+	InputTextFlagsCallbackCompletion = 1 << 6
+	// InputTextFlagsCallbackHistory for callback on pressing Up/Down arrows (for history handling).
+	InputTextFlagsCallbackHistory = 1 << 7
+	// InputTextFlagsCallbackAlways for callback on each iteration. User code may query cursor position, modify text buffer.
+	InputTextFlagsCallbackAlways = 1 << 8
+	// InputTextFlagsCallbackCharFilter for callback on character inputs to replace or discard them.
+	// Modify 'EventChar' to replace or discard, or return 1 in callback to discard.
+	InputTextFlagsCallbackCharFilter = 1 << 9
+	// InputTextFlagsAllowTabInput when pressing TAB to input a '\t' character into the text field.
+	InputTextFlagsAllowTabInput = 1 << 10
+	// InputTextFlagsCtrlEnterForNewLine in multi-line mode, unfocus with Enter, add new line with Ctrl+Enter
+	// (default is opposite: unfocus with Ctrl+Enter, add line with Enter).
+	InputTextFlagsCtrlEnterForNewLine = 1 << 11
+	// InputTextFlagsNoHorizontalScroll disables following the cursor horizontally.
+	InputTextFlagsNoHorizontalScroll = 1 << 12
+	// InputTextFlagsAlwaysInsertMode sets insert mode.
+	InputTextFlagsAlwaysInsertMode = 1 << 13
+	// InputTextFlagsReadOnly sets read-only mode.
+	InputTextFlagsReadOnly = 1 << 14
+	// InputTextFlagsPassword sets password mode, display all characters as '*'.
+	InputTextFlagsPassword = 1 << 15
+	// InputTextFlagsNoUndoRedo disables undo/redo. Note that input text owns the text data while active,
+	// if you want to provide your own undo/redo stack you need e.g. to call ClearActiveID().
+	InputTextFlagsNoUndoRedo = 1 << 16
+	// InputTextFlagsCharsScientific allows 0123456789.+-*/eE (Scientific notation input).
+	InputTextFlagsCharsScientific = 1 << 17
+	// inputTextFlagsCallbackResize for callback on buffer capacity change requests.
+	inputTextFlagsCallbackResize = 1 << 18
+)
+
 // InputTextV creates a text field for dynamic text input.
 //
 // Contrary to the original library, this wrapper does not limit the maximum number of possible characters.
@@ -314,6 +380,64 @@ func InputInt(label string, value *int32) bool {
 	return InputIntV(label, value, 1, 100, 0)
 }
 
+const (
+	// ColorEditFlagsNone default = 0
+	ColorEditFlagsNone = 0
+	// ColorEditFlagsNoAlpha ignores Alpha component (read 3 components from the input pointer).
+	ColorEditFlagsNoAlpha = 1 << 1
+	// ColorEditFlagsNoPicker disables picker when clicking on colored square.
+	ColorEditFlagsNoPicker = 1 << 2
+	// ColorEditFlagsNoOptions disables toggling options menu when right-clicking on inputs/small preview.
+	ColorEditFlagsNoOptions = 1 << 3
+	// ColorEditFlagsNoSmallPreview disables colored square preview next to the inputs. (e.g. to show only the inputs)
+	ColorEditFlagsNoSmallPreview = 1 << 4
+	// ColorEditFlagsNoInputs disables inputs sliders/text widgets (e.g. to show only the small preview colored square).
+	ColorEditFlagsNoInputs = 1 << 5
+	// ColorEditFlagsNoTooltip disables tooltip when hovering the preview.
+	ColorEditFlagsNoTooltip = 1 << 6
+	// ColorEditFlagsNoLabel disables display of inline text label (the label is still forwarded to the tooltip and picker).
+	ColorEditFlagsNoLabel = 1 << 7
+	// ColorEditFlagsNoSidePreview disables bigger color preview on right side of the picker, use small colored square preview instead.
+	ColorEditFlagsNoSidePreview = 1 << 8
+	// ColorEditFlagsNoDragDrop disables drag and drop target. ColorButton: disable drag and drop source.
+	ColorEditFlagsNoDragDrop = 1 << 9
+	// ColorEditFlagsNoBorder disables border (which is enforced by default)
+	ColorEditFlagsNoBorder = 1 << 10
+
+	// User Options (right-click on widget to change some of them). You can set application defaults using SetColorEditOptions().
+	// The idea is that you probably don't want to override them in most of your calls, let the user choose and/or call
+	// SetColorEditOptions() during startup.
+
+	// ColorEditFlagsAlphaBar shows vertical alpha bar/gradient in picker.
+	ColorEditFlagsAlphaBar = 1 << 16
+	// ColorEditFlagsAlphaPreview displays preview as a transparent color over a checkerboard, instead of opaque.
+	ColorEditFlagsAlphaPreview = 1 << 17
+	// ColorEditFlagsAlphaPreviewHalf displays half opaque / half checkerboard, instead of opaque.
+	ColorEditFlagsAlphaPreviewHalf = 1 << 18
+	// ColorEditFlagsHDR = (WIP) surrently only disable 0.0f..1.0f limits in RGBA edition.
+	// Note: you probably want to use ImGuiColorEditFlags_Float flag as well.
+	ColorEditFlagsHDR = 1 << 19
+	// ColorEditFlagsRGB sets the format as RGB
+	ColorEditFlagsRGB = 1 << 20
+	// ColorEditFlagsHSV sets the format as HSV
+	ColorEditFlagsHSV = 1 << 21
+	// ColorEditFlagsHEX sets the format as HEX
+	ColorEditFlagsHEX = 1 << 22
+	// ColorEditFlagsUint8 _display_ values formatted as 0..255.
+	ColorEditFlagsUint8 = 1 << 23
+	// ColorEditFlagsFloat _display_ values formatted as 0.0f..1.0f floats instead of 0..255 integers. No round-trip of value via integers.
+	ColorEditFlagsFloat = 1 << 24
+
+	// ColorEditFlagsPickerHueBar shows bar for Hue, rectangle for Sat/Value.
+	ColorEditFlagsPickerHueBar = 1 << 25
+	// ColorEditFlagsPickerHueWheel shows wheel for Hue, triangle for Sat/Value.
+	ColorEditFlagsPickerHueWheel = 1 << 26
+	// ColorEditFlagsInputRGB enables input and output data in RGB format.
+	ColorEditFlagsInputRGB = 1 << 27
+	// ColorEditFlagsInputHSV enables input and output data in HSV format.
+	ColorEditFlagsInputHSV = 1 << 28
+)
+
 // ColorEdit3 calls ColorEdit3V(label, col, 0)
 func ColorEdit3(label string, col *[3]float32) bool {
 	return ColorEdit3V(label, col, 0)
@@ -339,6 +463,56 @@ func ColorEdit4V(label string, col *[4]float32, flags int) bool {
 	ccol := (*C.float)(&col[0])
 	return C.iggColorEdit4(labelArg, ccol, C.int(flags)) != 0
 }
+
+const (
+	// ColorPickerFlagsNone default = 0
+	ColorPickerFlagsNone = 0
+	// ColorPickerFlagsNoPicker disables picker when clicking on colored square.
+	ColorPickerFlagsNoPicker = 1 << 2
+	// ColorPickerFlagsNoOptions disables toggling options menu when right-clicking on inputs/small preview.
+	ColorPickerFlagsNoOptions = 1 << 3
+	// ColorPickerFlagsNoAlpha ignoreс Alpha component (read 3 components from the input pointer).
+	ColorPickerFlagsNoAlpha = 1 << 1
+	// ColorPickerFlagsNoSmallPreview disables colored square preview next to the inputs. (e.g. to show only the inputs)
+	ColorPickerFlagsNoSmallPreview = 1 << 4
+	// ColorPickerFlagsNoInputs disables inputs sliders/text widgets (e.g. to show only the small preview colored square).
+	ColorPickerFlagsNoInputs = 1 << 5
+	// ColorPickerFlagsNoTooltip disables tooltip when hovering the preview.
+	ColorPickerFlagsNoTooltip = 1 << 6
+	// ColorPickerFlagsNoLabel disables display of inline text label (the label is still forwarded to the tooltip and picker).
+	ColorPickerFlagsNoLabel = 1 << 7
+	// ColorPickerFlagsNoSidePreview disables bigger color preview on right side of the picker, use small colored square preview instead.
+	ColorPickerFlagsNoSidePreview = 1 << 8
+
+	// User Options (right-click on widget to change some of them). You can set application defaults using SetColorEditOptions().
+	// The idea is that you probably don't want to override them in most of your calls, let the user choose and/or call
+	// SetColorPickerOptions() during startup.
+
+	// ColorPickerFlagsAlphaBar shows vertical alpha bar/gradient in picker.
+	ColorPickerFlagsAlphaBar = 1 << 16
+	// ColorPickerFlagsAlphaPreview displays preview as a transparent color over a checkerboard, instead of opaque.
+	ColorPickerFlagsAlphaPreview = 1 << 17
+	// ColorPickerFlagsAlphaPreviewHalf displays half opaque / half checkerboard, instead of opaque.
+	ColorPickerFlagsAlphaPreviewHalf = 1 << 18
+	// ColorPickerFlagsRGB sets the format as RGB
+	ColorPickerFlagsRGB = 1 << 20
+	// ColorPickerFlagsHSV sets the format as HSV
+	ColorPickerFlagsHSV = 1 << 21
+	// ColorPickerFlagsHEX sets the format as HEX
+	ColorPickerFlagsHEX = 1 << 22
+	// ColorPickerFlagsUint8 _display_ values formatted as 0..255.
+	ColorPickerFlagsUint8 = 1 << 23
+	// ColorPickerFlagsFloat _display_ values formatted as 0.0f..1.0f floats instead of 0..255 integers. No round-trip of value via integers.
+	ColorPickerFlagsFloat = 1 << 24
+	// ColorPickerFlagsPickerHueBar bar for Hue, rectangle for Sat/Value.
+	ColorPickerFlagsPickerHueBar = 1 << 25
+	// ColorPickerFlagsPickerHueWheel wheel for Hue, triangle for Sat/Value.
+	ColorPickerFlagsPickerHueWheel = 1 << 26
+	// ColorPickerFlagsInputRGB enables input and output data in RGB format.
+	ColorPickerFlagsInputRGB = 1 << 27
+	// ColorPickerFlagsInputHSV enables input and output data in HSV format.
+	ColorPickerFlagsInputHSV = 1 << 28
+)
 
 // ColorPicker3 calls ColorPicker3V(label, col, 0)
 func ColorPicker3(label string, col *[3]float32) bool {
@@ -373,6 +547,49 @@ func CollapsingHeader(label string) bool {
 	return C.iggCollapsingHeader(labelArg) != 0
 }
 
+const (
+	// TreeNodeFlagsNone default = 0
+	TreeNodeFlagsNone = 0
+	// TreeNodeFlagsSelected draws as selected.
+	TreeNodeFlagsSelected = 1 << 0
+	// TreeNodeFlagsFramed draws full colored frame (e.g. for CollapsingHeader).
+	TreeNodeFlagsFramed = 1 << 1
+	// TreeNodeFlagsAllowItemOverlap hit testing to allow subsequent widgets to overlap this one.
+	TreeNodeFlagsAllowItemOverlap = 1 << 2
+	// TreeNodeFlagsNoTreePushOnOpen doesn't do a TreePush() when open
+	// (e.g. for CollapsingHeader) = no extra indent nor pushing on ID stack.
+	TreeNodeFlagsNoTreePushOnOpen = 1 << 3
+	// TreeNodeFlagsNoAutoOpenOnLog doesn't automatically and temporarily open node when Logging is active
+	// (by default logging will automatically open tree nodes).
+	TreeNodeFlagsNoAutoOpenOnLog = 1 << 4
+	// TreeNodeFlagsDefaultOpen defaults node to be open.
+	TreeNodeFlagsDefaultOpen = 1 << 5
+	// TreeNodeFlagsOpenOnDoubleClick needs double-click to open node.
+	TreeNodeFlagsOpenOnDoubleClick = 1 << 6
+	// TreeNodeFlagsOpenOnArrow opens only when clicking on the arrow part.
+	// If TreeNodeFlagsOpenOnDoubleClick is also set, single-click arrow or double-click all box to open.
+	TreeNodeFlagsOpenOnArrow = 1 << 7
+	// TreeNodeFlagsLeaf allows no collapsing, no arrow (use as a convenience for leaf nodes).
+	TreeNodeFlagsLeaf = 1 << 8
+	// TreeNodeFlagsBullet displays a bullet instead of an arrow.
+	TreeNodeFlagsBullet = 1 << 9
+	// TreeNodeFlagsFramePadding uses FramePadding (even for an unframed text node) to
+	// vertically align text baseline to regular widget height. Equivalent to calling AlignTextToFramePadding().
+	TreeNodeFlagsFramePadding = 1 << 10
+	// TreeNodeFlagsSpanAvailWidth extends hit box to the right-most edge, even if not framed.
+	// This is not the default in order to allow adding other items on the same line.
+	// In the future we may refactor the hit system to be front-to-back, allowing natural overlaps
+	// and then this can become the default.
+	TreeNodeFlagsSpanAvailWidth = 1 << 11
+	// TreeNodeFlagsSpanFullWidth extends hit box to the left-most and right-most edges (bypass the indented area).
+	TreeNodeFlagsSpanFullWidth = 1 << 12
+	// TreeNodeFlagsNavLeftJumpsBackHere (WIP) Nav: left direction may move to this TreeNode() from any of its child
+	// (items submitted between TreeNode and TreePop)
+	TreeNodeFlagsNavLeftJumpsBackHere = 1 << 13
+	// TreeNodeFlagsCollapsingHeader combines TreeNodeFlagsFramed and TreeNodeFlagsNoAutoOpenOnLog.
+	TreeNodeFlagsCollapsingHeader = TreeNodeFlagsFramed | TreeNodeFlagsNoTreePushOnOpen | TreeNodeFlagsNoAutoOpenOnLog
+)
+
 // TreeNodeV returns true if the tree branch is to be rendered. Call TreePop() in this case.
 func TreeNodeV(label string, flags int) bool {
 	labelArg, labelFin := wrapString(label)
@@ -399,6 +616,19 @@ func SetNextItemOpen(open bool, cond Condition) {
 func TreeNodeToLabelSpacing() float32 {
 	return float32(C.iggGetTreeNodeToLabelSpacing())
 }
+
+const (
+	// SelectableFlagsNone default = 0
+	SelectableFlagsNone = 0
+	// SelectableFlagsDontClosePopups makes clicking the selectable not close any parent popup windows.
+	SelectableFlagsDontClosePopups = 1 << 0
+	// SelectableFlagsSpanAllColumns allows the selectable frame to span all columns (text will still fit in current column).
+	SelectableFlagsSpanAllColumns = 1 << 1
+	// SelectableFlagsAllowDoubleClick generates press events on double clicks too.
+	SelectableFlagsAllowDoubleClick = 1 << 2
+	// SelectableFlagsDisabled disallows selection and displays text in a greyed out color.
+	SelectableFlagsDisabled = 1 << 3
+)
 
 // SelectableV returns true if the user clicked it, so you can modify your selection state.
 // flags are the SelectableFlags to apply.
@@ -656,6 +886,35 @@ func ColumnsCount() int {
 	return int(C.iggGetColumnsCount())
 }
 
+const (
+	// TabBarFlagsNone default = 0.
+	TabBarFlagsNone = 0
+	// TabBarFlagsReorderable Allow manually dragging tabs to re-order them + New tabs are appended at the end of list
+	TabBarFlagsReorderable = 1 << 0
+	// TabBarFlagsAutoSelectNewTabs Automatically select new tabs when they appear
+	TabBarFlagsAutoSelectNewTabs = 1 << 1
+	// TabBarFlagsTabListPopupButton Disable buttons to open the tab list popup
+	TabBarFlagsTabListPopupButton = 1 << 2
+	// TabBarFlagsNoCloseWithMiddleMouseButton Disable behavior of closing tabs (that are submitted with p_open != NULL)
+	// with middle mouse button. You can still repro this behavior on user's side with if
+	// (IsItemHovered() && IsMouseClicked(2)) *p_open = false.
+	TabBarFlagsNoCloseWithMiddleMouseButton = 1 << 3
+	// TabBarFlagsNoTabListScrollingButtons Disable scrolling buttons (apply when fitting policy is
+	// TabBarFlagsFittingPolicyScroll)
+	TabBarFlagsNoTabListScrollingButtons = 1 << 4
+	// TabBarFlagsNoTooltip Disable tooltips when hovering a tab
+	TabBarFlagsNoTooltip = 1 << 5
+	// TabBarFlagsFittingPolicyResizeDown Resize tabs when they don't fit
+	TabBarFlagsFittingPolicyResizeDown = 1 << 6
+	// TabBarFlagsFittingPolicyScroll Add scroll buttons when tabs don't fit
+	TabBarFlagsFittingPolicyScroll = 1 << 7
+	// TabBarFlagsFittingPolicyMask combines
+	// TabBarFlagsFittingPolicyResizeDown and TabBarFlagsFittingPolicyScroll
+	TabBarFlagsFittingPolicyMask = TabBarFlagsFittingPolicyResizeDown | TabBarFlagsFittingPolicyScroll
+	// TabBarFlagsFittingPolicyDefault alias for TabBarFlagsFittingPolicyResizeDown
+	TabBarFlagsFittingPolicyDefault = TabBarFlagsFittingPolicyResizeDown
+)
+
 // BeginTabBarV create and append into a TabBar
 func BeginTabBarV(strID string, flags int) bool {
 	idArg, idFin := wrapString(strID)
@@ -673,6 +932,23 @@ func BeginTabBar(strID string) bool {
 func EndTabBar() {
 	C.iggEndTabBar()
 }
+
+const (
+	// TabItemFlagsNone default = 0
+	TabItemFlagsNone = 0
+	// TabItemFlagsUnsavedDocument Append '*' to title without affecting the ID, as a convenience to avoid using the
+	// ### operator. Also: tab is selected on closure and closure is deferred by one frame to allow code to undo it
+	// without flicker.
+	TabItemFlagsUnsavedDocument = 1 << 0
+	// TabItemFlagsSetSelected Trigger flag to programmatically make the tab selected when calling BeginTabItem()
+	TabItemFlagsSetSelected = 1 << 1
+	// TabItemFlagsNoCloseWithMiddleMouseButton  Disable behavior of closing tabs (that are submitted with
+	// p_open != NULL) with middle mouse button. You can still repro this behavior on user's side with if
+	// (IsItemHovered() && IsMouseClicked(2)) *p_open = false.
+	TabItemFlagsNoCloseWithMiddleMouseButton = 1 << 2
+	// TabItemFlagsNoPushID Don't call PushID(tab->ID)/PopID() on BeginTabItem()/EndTabItem()
+	TabItemFlagsNoPushID = 1 << 3
+)
 
 // BeginTabItemV create a Tab. Returns true if the Tab is selected.
 func BeginTabItemV(label string, open *bool, flags int) bool {
